mono-designer/pyproject.toml
[project]
name = "mono-designer"
version = "0.3.0"
description = "Monospace Design TUI - Designer Toolset"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0.0",
    "PyYAML>=5.4.1",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "jsonschema>=4.0.0",
]

[project.optional-dependencies]
dev = ["pytest"]

[project.scripts]
mono-designer = "mono_designer.cli.main:app"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["mono_designer*"]
